refactor(cli): extract check cluster into commands/check.ts (5.0-alpha)
Eighth cut of the cli.ts god-module split — first of the previously-deferred
entangled chain (check -> design -> install -> login -> setup Phase B + review).

- New src/commands/check.ts: cmdCheck (exported COMMANDS entry) + its
  module-private sub-handler cmdVerifyAttestation (routed via `kit check
  verify-attestation`). Both were check-only.
- cmdCheck imports KIT_VERSION + JsonCheck/JsonCheckOutput/autoInstallScanners/
  maybeEmitCheckAttestation from the neutral cli-checks-shared module, and
  selfUpgrade from the upgrade command module — no cli.ts back-import.
- cli.ts imports cmdCheck back; the still-inline orchestrators cmdReview +
  cmdSetup now call the imported cmdCheck. COMMANDS table unchanged.
- Prunes now check-only imports from cli.ts (findings-track, check-web-search,
  check-lock, output table printers, checkHooks, hints, envTruthy, selfUpgrade,
  autoInstallScanners, maybeEmitCheckAttestation); keeps symbols still used by
  other handlers (runStep, isGitRepository, checkServices/Secrets/Skills, the
  ci/self-audit shared symbols).
- Dynamic imports re-pathed ./ -> ../.

cli.ts: 3613 -> 3219 lines (-394; -3111 cumulative from the original 6330,
~49% smaller). Typecheck, eslint (0 errors), full suite (2580), prettier all
green; kit check --json + verify-attestation smoke-tested.
